healthpoints/requirements/requirements-dev.txt
-r requirements.txt

django-extensions
django-debug-toolbar
werkzeug
ipython
ipdb